move python dir to proper location
