+++ python/tflite_micro/signal/__init__.py
@@ -1,0 +1,1 @@
+# Empty file required by setuptools.find_packages to recognize this as a package
