--- signal/python/ops/__init__.py
@@ -1,1 +1,0 @@
-# Empty file required by setuptools.find_packages to recognize this as a package
