--- tensorflow/lite/micro/tools/make/additional_kernels.inc
@@ -1,1 +1,0 @@
-include $(TENSORFLOW_ROOT)signal/micro/kernels/Makefile_kernels.inc
