--- tensorflow/lite/micro/tools/make/additional_tests.inc
@@ -1,1 +1,0 @@
-include $(TENSORFLOW_ROOT)signal/micro/kernels/Makefile_tests.inc
